requirements.txt
alembic==0.9.9
asn1crypto==0.24.0
astroid==1.6.3
attrs==17.4.0
autopep8==1.3.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.0.3
coveralls==1.3.0
docopt==0.6.2
Flask==0.12.2
gunicorn==19.7.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
pipenv==11.8.0
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.2
pycodestyle==2.3.1
pycparser==2.18
PyJWT==1.6.1
pytest==3.4.1
pytest-ordering==0.5
python-dateutil==2.7.2
python-editor==1.0.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.5
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
wrapt==1.10.11
